re-order some files fro clarity
